Delete more options on uninstall

+++ uninstall.php
@@ -10,11 +10,12 @@ defined( 'WP_UNINSTALL_PLUGIN' ) || exit;
 
 $options = get_option( 'ang_options' );
 
-if ( is_array( $options ) ) {
-	if ( isset( $options['remove_on_uninstall'] ) && true === $options['remove_on_uninstall'] ) {
-		delete_option( 'ang_options' );
-	}
+if ( is_array( $options ) && isset( $options['remove_on_uninstall'] ) && true === $options['remove_on_uninstall'] ) {
+	delete_option( 'ang_options' );
+	delete_option( '_ang_import_history' );
 }
+
+delete_option( '_ang_installed_time' );
 
 delete_transient( 'ang_license_message' );
 delete_transient( 'analogwp_template_info' );
